+++ test/flutter_crisp_chat_test.dart
@@ -1,7 +1,7 @@
-import 'package:flutter_test/flutter_test.dart';
 import 'package:crisp_chat/crisp_chat.dart';
-import 'package:crisp_chat/src/flutter_crisp_chat_platform_interface.dart';
 import 'package:crisp_chat/src/flutter_crisp_chat_method_channel.dart';
+import 'package:crisp_chat/src/flutter_crisp_chat_platform_interface.dart';
+import 'package:flutter_test/flutter_test.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 /// [MockFlutterCrispChatPlatform] is for testing method channel call.
@@ -27,6 +27,11 @@ class MockFlutterCrispChatPlatform
   @override
   Future<void> setSessionInt({required String key, required int value}) {
     //
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<String?> getSessionIdentifier() {
     throw UnimplementedError();
   }
 }